% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nprobw}
\alias{nprobw}
\title{Summarizing the sample size estimation.}
\usage{
nprobw(x, n)
}
\arguments{
\item{x}{an object class of "RDSsamplesize", results of estimated sample size distribution of a call to 'calSize'.}

\item{n}{integer; target sample size.}
}
\value{
a table presenting the probability of the accumulated sample size (including seeds) reaching at least \emph{n}
by each wave, w=1,..., \emph{maxWave}
}
\description{
Summarizing the sample size estimation.
}
\examples{
x <- calSize(s=10,c=3,maxWave=9,rr=0.3,bruteMC=FALSE,tol=0.025)
nprobw(x,n=100)
}
